[Style] 뮤지컬관련정보 컴포넌트 분리 및 로더임시적용
